+++ aderyn_core/src/ast/ast_nodes.rs
@@ -66,3 +66,18 @@ pub struct ContractDefinition {
     pub src: String,
     pub id: NodeID,
 }
+
+#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
+#[serde(untagged)]
+pub enum Documentation {
+    String(Option<String>),
+    Structured(Option<StructuredDocumentation>),
+}
+
+#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
+#[serde(rename_all = "camelCase")]
+pub struct StructuredDocumentation {
+    pub text: String,
+    pub src: String,
+    pub id: NodeID,
+}
