+++ aderyn_core/src/ast/impls/node/documentation.rs
@@ -1,14 +1,5 @@
-use super::*;
-use crate::visitor::ast_visitor::*;
+use crate::{ast::*, visitor::ast_visitor::*};
 use eyre::Result;
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(untagged)]
-pub enum Documentation {
-    String(Option<String>),
-    Structured(Option<StructuredDocumentation>),
-}
 
 impl Node for Documentation {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
@@ -28,14 +19,6 @@ impl Node for Documentation {
             }
         }
     }
-}
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct StructuredDocumentation {
-    pub text: String,
-    pub src: String,
-    pub id: NodeID,
 }
 
 impl Node for StructuredDocumentation {
