+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -1,2 +1,3 @@
 pub mod blocks;
 pub mod contracts;
+pub mod documentation;
