+++ aderyn_core/src/ast/mod.rs
@@ -1,5 +1,4 @@
 pub mod ast_nodes;
-pub mod documentation;
 pub mod enumerations;
 pub mod errors;
 pub mod events;
@@ -24,10 +23,10 @@ pub mod variables;
 pub mod yul;
 
 pub use self::{
-    documentation::*, enumerations::*, errors::*, events::*, expressions::*, functions::*,
-    identifiers::*, impls::*, import_directives::*, literals::*, magic::*, modifiers::*,
-    node_type::*, pragma_directives::*, source_units::*, statements::*, structures::*, types::*,
-    user_defined_value_types::*, using_for_directives::*, variables::*, yul::*,
+    enumerations::*, errors::*, events::*, expressions::*, functions::*, identifiers::*, impls::*,
+    import_directives::*, literals::*, magic::*, modifiers::*, node_type::*, pragma_directives::*,
+    source_units::*, statements::*, structures::*, types::*, user_defined_value_types::*,
+    using_for_directives::*, variables::*, yul::*,
 };
 
 pub use ast_nodes::*;
